requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
protobuf==3.20.3